Module 4 Project Files
Updated the website to include all pages for the website. Fixed the database functionality for the employees page so that it will now always display the information. Added a new database for the menu page that displays some items and their prices. Added a contact form to the contact page.